+++ src/ts/Ads/Views/VideoOverlay.ts
@@ -64,6 +64,7 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
     protected _campaign: Campaign;
 
     private _useCloseIconInsteadOfSkipIcon: boolean | undefined = false;
+    private _disableFadeOutOnClick: boolean | undefined = false;
 
     constructor(parameters: IVideoOverlayParameters<Campaign>, privacy: AbstractPrivacy, showGDPRBanner: boolean, showPrivacyDuringVideo: boolean) {
         super(parameters.platform, 'video-overlay', parameters.placement.muteVideo());
@@ -77,6 +78,9 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
         this._showGDPRBanner = showGDPRBanner;
         this._showPrivacyDuringVideo = showPrivacyDuringVideo;
         this._useCloseIconInsteadOfSkipIcon = parameters.placement.useCloseIconInsteadOfSkipIcon();
+
+        //Disable click fadeout for placements that disabled overlay fadeout for Mobilityware
+        this._disableFadeOutOnClick = CustomFeatures.shouldVideoOverlayRemainVisible(this._gameId) && parameters.placement.disableVideoControlsFade();
 
         this._templateData = {
             muted: parameters.placement.muteVideo()
@@ -367,6 +371,10 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
     }
 
     private onClick(event: Event) {
+        if (this._disableFadeOutOnClick) {
+            return;
+        }
+
         this.resetFadeTimer();
 
         if (this._areControlsVisible) {